requirements.txt
customtkinter>=5.2.0
ffmpeg-python>=0.2.0
mutagen>=1.47.0
pillow>=10.0.0